added type dependency
