+++ cuda/lib/types.cpp
@@ -1,7 +1,9 @@
 #ifndef _TYPES
 #define _TYPES
 
+#include <unordered_map>
 #include <string>
+#include <vector>
 using namespace std;
 
 typedef unordered_map< string, vector<float> > data_map;
