+++ cuda/main.cu
@@ -1,4 +1,5 @@
 #include "./lib/read_csv.cpp"
+#include "../lib/types.cpp"
 using namespace std;
 
 void print_cluster(int k, int *cluster_assignment,int n,const college_dataset& data){
@@ -24,28 +25,6 @@ void print_cluster_size(int k, int *cluster_assignment,int n){
   for (int i=0; i<k; i++){
     cout << i << " ::: " << size[i] << endl;
   }
-}
-
-college_dataset fill_college_struct(const data_map& college_data){
-
-    vector<float> sample_map_data = college_data.begin()->second; 
-  
-    college_dataset data;
-    data.size = college_data.size();
-    data.dimensions = sample_map_data.size();
-    data.names = new string[data.size];
-    data.features = new float *[data.size];
-    for (int i=0; i<data.size; i++)
-      data.features[i] = new float[data.dimensions];
-  
-    int index = 0;
-    for (data_map::const_iterator it = college_data.begin(); it != college_data.end(); it++) {
-        data.names[index] = it->first;
-        for (int j=0; j<data.dimensions; j++)
-          data.features[index][j] = (it->second)[j];
-        index++;
-    } 
-    return data;
 }
 
 
@@ -147,10 +126,8 @@ for (int i=0; i<data.size; i++){
 }
 
 
-
-
 int * find_clusters(int k, const college_dataset data, int max_iter) {
-int iter = 0;
+// int iter = 0;
 int * cluster_size;
 int * cluster_assignment;
